feat: credit adapter

+++ src/interfaces/IWmatic.sol
@@ -2,12 +2,9 @@
 pragma solidity ^0.8.0;
 
 interface IWmatic {
-    function allowance(
-        address spender,
-        address caller
-    ) external returns (uint256);
+    function allowance(address spender, address caller) external returns (uint256);
 
     function deposit() external payable;
 
-    function approve(address guy, uint wad) external returns (bool);
+    function approve(address guy, uint256 wad) external returns (bool);
 }
